# 218 json model element performance