# Make sure these boxes are signed before submitting your Pull Request -- thank you.

# Must haves
- [x] I have read and followed the contributing guide lines at https://github.com/ait-aecid/logdata-anomaly-miner/wiki/Git-development-workflow
- [x] Issues exist for this PR
- [x] I added related issues using the "Fixes #<issue-id>"-notations
- [x] This Pull-Requests merges into the "development"-branch

Fixes #1149 

# Submission specific

- [ ] This PR introduces breaking changes
- [ ] My change requires a change to the documentation
- [ ] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] All new and existing tests passed

# Describe changes:
There have been quite some improvements on the performance of the JsonModelElement.
Before improvements:

waiting time: 1 seconds
The JsonModelElement could in average handle 265 LogAtoms [160, 238, 210, 288, 291, 303, 254, 296, 305, 304]
Ran 1 test in 93.831s
OK

----------------------------------------------------------------------
Improvements initial commit:
waiting time: 1 seconds
The JsonModelElement could in average handle 283 LogAtoms [287, 293, 253, 304, 305, 306, 251, 303, 244, 286]
Ran 1 test in 83.095s
OK

----------------------------------------------------------------------
Improvements in is_escaped_unicode:
waiting time: 1 seconds
The JsonModelElement could in average handle 587 LogAtoms [511, 616, 626, 585, 634, 587, 408, 632, 633, 637]
Ran 1 test in 45.786s
OK

----------------------------------------------------------------------
Small Improvements in code:
waiting time: 1 seconds
The JsonModelElement could in average handle 606 LogAtoms [622, 662, 644, 646, 632, 668, 515, 483, 598, 593]
Ran 1 test in 42.837s
OK